+++ C/VCBounCballs/BounCballs/.vs/BounCballs/v17/DocumentLayout.json
@@ -1,0 +1,41 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "D:\\BouncyBalls\\BouncyBalls\\C\\VCBounCballs\\BounCballs\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{9A5CB312-1A56-4100-905D-00F59F99ECB5}|BounCballs\\BounCballs.vcxproj|D:\\BouncyBalls\\BouncyBalls\\C\\VCBounCballs\\BounCballs\\BounCballs\\main.c||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
+      "RelativeMoniker": "D:0:0:{9A5CB312-1A56-4100-905D-00F59F99ECB5}|BounCballs\\BounCballs.vcxproj|solutionrelative:BounCballs\\main.c||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 200,
+          "SelectedChildIndex": 1,
+          "Children": [
+            {
+              "$type": "Bookmark",
+              "Name": "ST:0:0:{cce594b6-0c39-4442-ba28-10c64ac7e89f}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "main.c",
+              "DocumentMoniker": "D:\\BouncyBalls\\BouncyBalls\\C\\VCBounCballs\\BounCballs\\BounCballs\\main.c",
+              "RelativeDocumentMoniker": "BounCballs\\main.c",
+              "ToolTip": "D:\\BouncyBalls\\BouncyBalls\\C\\VCBounCballs\\BounCballs\\BounCballs\\main.c",
+              "RelativeToolTip": "BounCballs\\main.c",
+              "ViewState": "AQIAACoAAAAAAAAAAAAAADoAAAAbAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000423|",
+              "WhenOpened": "2024-08-13T17:24:18.013Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
