+++ C/VCBounCballs/BounCballs/BounCballs.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35027.167
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BounCballs", "BounCballs\BounCballs.vcxproj", "{9A5CB312-1A56-4100-905D-00F59F99ECB5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Debug|x64.ActiveCfg = Debug|x64
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Debug|x64.Build.0 = Debug|x64
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Debug|x86.ActiveCfg = Debug|Win32
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Debug|x86.Build.0 = Debug|Win32
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Release|x64.ActiveCfg = Release|x64
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Release|x64.Build.0 = Release|x64
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Release|x86.ActiveCfg = Release|Win32
+		{9A5CB312-1A56-4100-905D-00F59F99ECB5}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {68322CD0-1E0C-4AD5-8CC0-C2F878BCBB2F}
+	EndGlobalSection
+EndGlobal
