+++ C/VCBounCballs/BounCballs/BounCballs/x64/Debug/BounCballs.exe.recipe
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project>
+  <ProjectOutputs>
+    <ProjectOutput>
+      <FullPath>D:\BouncyBalls\BouncyBalls\C\VCBounCballs\BounCballs\x64\Debug\BounCballs.exe</FullPath>
+    </ProjectOutput>
+  </ProjectOutputs>
+  <ContentFiles />
+  <SatelliteDlls />
+  <NonRecipeFileRefs />
+</Project>
