+++ C/VCBounCballs/BounCballs/BounCballs/x64/Debug/BounCballs.log
@@ -1,0 +1,1 @@
+﻿  BounCballs.vcxproj -> D:\BouncyBalls\BouncyBalls\C\VCBounCballs\BounCballs\x64\Debug\BounCballs.exe
