+++ C/VCBounCballs/BounCballs/BounCballs/x64/Debug/BounCballs.tlog/BounCballs.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.40.33807:TargetPlatformVersion=10.0.26100.0:
+Debug|x64|D:\BouncyBalls\BouncyBalls\C\VCBounCballs\BounCballs\|
