+++ C/VCBounCballs/BounCballs/BounCballs/x64/Debug/BounCballs.tlog/Cl.items.tlog
@@ -1,0 +1,1 @@
+D:\BouncyBalls\BouncyBalls\C\VCBounCballs\BounCballs\BounCballs\main.c;D:\BouncyBalls\BouncyBalls\C\VCBounCballs\BounCballs\BounCballs\x64\Debug\main.obj
